Replace Diesel with raw SQL

+++ migrations/2018-02-18-063750_create_chests/up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE chests (
   id       INTEGER PRIMARY KEY NOT NULL,
-  position BIGINT              NOT NULL,
-  lv       SMALLINT            NOT NULL,
-  found_by BIGINT
+  position INTEGER             NOT NULL,
+  lv       INTEGER             NOT NULL,
+  found_by INTEGER
 )
