+++ migrations/2018-03-11-150821_create_users/up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE users (
-  id       BIGINT PRIMARY KEY NOT NULL,
-  uuid     TEXT               NOT NULL,
-  token    BIGINT
+  id       INTEGER PRIMARY KEY NOT NULL,
+  uuid     TEXT                NOT NULL,
+  token    INTEGER
 )
